Rework/formalize printer calibration
This adds a separate action to generate a calibration PDF (with
instructions!) and also adds calibration bars to the main output PDFs
just for double checking.

The main difference between the two is that the bars in the calibration
PDF are not rescaled with the calibration factors, so they can be used
in a one-off step to set relative factors. The bars in the main PDF are
rescaled, which means that they can function as a cross-check on wether
the calibration actually worked.

The calibration factors live in the js/pdf.js code. That's a bit
unfriendly to random users, but anything that would live in the page
itself would be too ephemeral to be nice to use as well. And hopefully
people just have to do this once for their particular OS/drivers/printer
combo anyway!

This is, as usual, not impressively elegantly implemented. But it should
be very straightforward at least.

Major style malus for the way we're loading the image for the
calibration bars, though. Sorry.