Add water/mountain map gen
All of this code is pretty poor, sorry.
This adds generating water and mountain tiles. The algorithm attempts to
generate more linear runs of those than the "regular" tile types. Not
working super beautifully, but probably good enough for a part of the
implementation that's kind of best effort. (Did I mention that none of
this map generation guarantees that it will actually terminate?)